requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
anthropic>=0.5.0
python-dotenv>=0.19.0
pydantic>=1.8.0
python-multipart>=0.0.5
Jinja2>=3.0.0
